+++ zcash_primitives/src/constants/testnet.rs
@@ -5,7 +5,7 @@
 /// [SLIP 44]: https://github.com/satoshilabs/slips/blob/master/slip-0044.md
 pub const COIN_TYPE: u32 = 1;
 
-/// The HRP for a Bech32-encoded testnet [`ExtendedSpendingKey`].
+/// The HRP for a Bech32-encoded testnet Sapling [`ExtendedSpendingKey`].
 ///
 /// Defined in [ZIP 32].
 ///
@@ -13,7 +13,7 @@ pub const COIN_TYPE: u32 = 1;
 /// [ZIP 32]: https://github.com/zcash/zips/blob/master/zip-0032.rst
 pub const HRP_SAPLING_EXTENDED_SPENDING_KEY: &str = "secret-extended-key-test";
 
-/// The HRP for a Bech32-encoded testnet [`ExtendedFullViewingKey`].
+/// The HRP for a Bech32-encoded testnet Sapling [`ExtendedFullViewingKey`].
 ///
 /// Defined in [ZIP 32].
 ///
@@ -21,7 +21,7 @@ pub const HRP_SAPLING_EXTENDED_SPENDING_KEY: &str = "secret-extended-key-test";
 /// [ZIP 32]: https://github.com/zcash/zips/blob/master/zip-0032.rst
 pub const HRP_SAPLING_EXTENDED_FULL_VIEWING_KEY: &str = "zxviewtestsapling";
 
-/// The HRP for a Bech32-encoded testnet [`PaymentAddress`].
+/// The HRP for a Bech32-encoded testnet Sapling [`PaymentAddress`].
 ///
 /// Defined in section 5.6.4 of the [Zcash Protocol Specification].
 ///
@@ -29,12 +29,12 @@ pub const HRP_SAPLING_EXTENDED_FULL_VIEWING_KEY: &str = "zxviewtestsapling";
 /// [Zcash Protocol Specification]: https://github.com/zcash/zips/blob/master/protocol/protocol.pdf
 pub const HRP_SAPLING_PAYMENT_ADDRESS: &str = "ztestsapling";
 
-/// The prefix for a Base58Check-encoded testnet [`TransparentAddress::PublicKey`].
+/// The prefix for a Base58Check-encoded testnet transparent [`PublicKeyHash`].
 ///
-/// [`TransparentAddress::PublicKey`]: crate::legacy::TransparentAddress::PublicKey
+/// [`PublicKeyHash`]: crate::legacy::TransparentAddress::PublicKeyHash
 pub const B58_PUBKEY_ADDRESS_PREFIX: [u8; 2] = [0x1d, 0x25];
 
-/// The prefix for a Base58Check-encoded testnet [`TransparentAddress::Script`].
+/// The prefix for a Base58Check-encoded testnet transparent [`ScriptHash`].
 ///
-/// [`TransparentAddress::Script`]: crate::legacy::TransparentAddress::Script
+/// [`ScriptHash`]: crate::legacy::TransparentAddress::ScriptHash
 pub const B58_SCRIPT_ADDRESS_PREFIX: [u8; 2] = [0x1c, 0xba];