+++ zcash_primitives/CHANGELOG.md
@@ -64,6 +64,9 @@ and this library adheres to Rust's notion of
     defaults to `constants::regtest::` for everything else. 
 
 ### Changed
+- `zcash_primitives::legacy::TransparentAddress` variants have changed:
+  - `TransparentAddress::PublicKey` has been renamed to `PublicKeyHash`
+  - `TransparentAddress::Script` has been renamed to `ScriptHash`
 - `zcash_primitives::transaction`:
   - `builder::Builder` now has a generic parameter for the type of progress
     notifier, which needs to implement `sapling::builder::ProverProgress` in